Rename server.rs to core.rs
